qcacld: reset public wifi driver to Release 5.2.052.2D
CAF used to push tags ahead of the SoCs release a few years back. This changed a few years back after several unintended leaks caused by this practice.

Unfortunately OnePlus does not release the WiFi-Source as it is not licensed under GPL.
Using the oldest available tag for WiFi driver will result in non working VoWifi functionality on the OnePlus 8 Pro.

Resetting the WiFi driver to an older revision fixes this issue.